+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Only latest stop reason should be printed to logs on startup
 - Fix bug where geofencing client wasn't initialized properly, leading to very unreliable region transition detection
 - Fix bug where some settings (`pubQos`, `mqttProtocolLevel` etc.) couldn't be set via the config editor (#1801)
+- Fix crash when trying to decode an invalid face image on an info card
 
 ## Version 2.5.1